Finalized attribute_assert.
